requirements.txt
yfinance
dotenv
pandas
numpy
scikit-learn
sqlalchemy
pathlib
psycopg2-binary
tensorflow
fastapi
uvicorn
pytest